Write up
I went through and started a quick write up explaining each of the
different datasets.  I then wrote them to a csv so we always have
something easy to grab when doing the write up.

The only problem I had was with foodstamps and how we would want to
display that data for the project as in shiny it would be different
datasets.

I also can’t actually run knitr on my computer since I don’t have tex
sooooo can you check and make sure it will compile haha.